saving progress on FOL_Build
