--- Scrub_Team_2018/FOL_Output/GAZ_output.txt
@@ -1,33 +1,0 @@
-%  http://purl.obolibrary.org/obo/GAZ_00000448
-%  source ontology: GAZ
-%  entity: http://purl.obolibrary.org/obo/GAZ_00000448
-%  label: jurisdiction
-%  parent: http://purl.obolibrary.org/obo/BFO_0000029
-%  parentLabel: site
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (jurisdiction(x) -> site(x)). 
-
-%  http://purl.obolibrary.org/obo/GAZ_00000448
-%  source ontology: GAZ
-%  entity: http://purl.obolibrary.org/obo/GAZ_00000448
-%  label: geographic location
-%  parent: http://purl.obolibrary.org/obo/BFO_0000029
-%  parentLabel: site
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (geographic_location(x) -> site(x)). 
-
-%  http://purl.obolibrary.org/obo/GAZ_00000448
-%  source ontology: GAZ
-%  entity: http://purl.obolibrary.org/obo/GAZ_00000448
-%  label: repository of documents
-%  parent: http://purl.obolibrary.org/obo/BFO_0000029
-%  parentLabel: site
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (repository_of_documents(x) -> site(x)). 
-
