--- Scrub_Team_2018/FOL_Output/NCBITaxon_output.txt
@@ -1,44 +1,0 @@
-%  http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  source ontology: NCBITaxon
-%  entity: http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  label: Homo sapiens
-%  parent: http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  parentLabel: Homo sapiens
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (Homo_sapiens(x) -> Homo_sapiens(x)). 
-
-%  http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  source ontology: NCBITaxon
-%  entity: http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  label: Homo sapiens
-%  parent: http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  parentLabel: Homo sapiens
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (Homo_sapiens(x) -> Homo_sapiens(x)). 
-
-%  http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  source ontology: NCBITaxon
-%  entity: http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  label: Homo sapiens
-%  parent: http://purl.obolibrary.org/obo/OBI_0100026
-%  parentLabel: organism
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (Homo_sapiens(x) -> organism(x)). 
-
-%  http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  source ontology: NCBITaxon
-%  entity: http://purl.obolibrary.org/obo/NCBITaxon_9606
-%  label: Homo sapiens
-%  parent: http://purl.obolibrary.org/obo/OBI_0100026
-%  parentLabel: organism
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (Homo_sapiens(x) -> organism(x)). 
-
