--- Scrub_Team_2018/FOL_Output/OMIABIS_output.txt
@@ -1,88 +1,0 @@
-%  http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  source ontology: OMIABIS
-%  entity: http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  label: patent
-%  parent: http://purl.obolibrary.org/obo/IAO_0000310
-%  parentLabel: document
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (patent(x) -> document(x)). 
-
-%  http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  source ontology: OMIABIS
-%  entity: http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  label: medical record
-%  parent: http://purl.obolibrary.org/obo/IAO_0000310
-%  parentLabel: document
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (medical_record(x) -> document(x)). 
-
-%  http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  source ontology: OMIABIS
-%  entity: http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  label: publication
-%  parent: http://purl.obolibrary.org/obo/IAO_0000310
-%  parentLabel: document
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (publication(x) -> document(x)). 
-
-%  http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  source ontology: OMIABIS
-%  entity: http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  label: regulatory document
-%  parent: http://purl.obolibrary.org/obo/IAO_0000310
-%  parentLabel: document
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (regulatory_document(x) -> document(x)). 
-
-%  http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  source ontology: OMIABIS
-%  entity: http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  label: informed assent form
-%  parent: http://purl.obolibrary.org/obo/IAO_0000310
-%  parentLabel: document
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (informed_assent_form(x) -> document(x)). 
-
-%  http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  source ontology: OMIABIS
-%  entity: http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  label: report
-%  parent: http://purl.obolibrary.org/obo/IAO_0000310
-%  parentLabel: document
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (report(x) -> document(x)). 
-
-%  http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  source ontology: OMIABIS
-%  entity: http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  label: questionnaire
-%  parent: http://purl.obolibrary.org/obo/IAO_0000310
-%  parentLabel: document
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (questionnaire(x) -> document(x)). 
-
-%  http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  source ontology: OMIABIS
-%  entity: http://purl.obolibrary.org/obo/ico.owl/http://purl.obolibrary.org/obo/OMIABIS_0001026
-%  label: document act input document
-%  parent: http://purl.obolibrary.org/obo/IAO_0000310
-%  parentLabel: document
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (document_act_input_document(x) -> document(x)). 
-
