--- Scrub_Team_2018/FOL_Output/PATO_output.txt
@@ -1,154 +1,0 @@
-%  http://purl.obolibrary.org/obo/PATO_0000125
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000125
-%  label: length
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (length(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000125
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000125
-%  label: relational quality
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (relational_quality(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000125
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000125
-%  label: information carrier
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: Na4befbdf021a4f74b92c64fc324e7ed3
-%  SubclassAxiom: Na4befbdf021a4f74b92c64fc324e7ed3
-
-all x (information_carrier(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000125
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000125
-%  label: mechanical wave quality
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (mechanical_wave_quality(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000125
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000125
-%  label: writing quality
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (writing_quality(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000125
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000125
-%  label: mass
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (mass(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000125
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000125
-%  label: pressure
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (pressure(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000122
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000122
-%  label: length
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (length(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000122
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000122
-%  label: relational quality
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (relational_quality(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000122
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000122
-%  label: information carrier
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: Na4befbdf021a4f74b92c64fc324e7ed3
-%  SubclassAxiom: Na4befbdf021a4f74b92c64fc324e7ed3
-
-all x (information_carrier(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000122
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000122
-%  label: mechanical wave quality
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (mechanical_wave_quality(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000122
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000122
-%  label: writing quality
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (writing_quality(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000122
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000122
-%  label: mass
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (mass(x) -> quality(x)). 
-
-%  http://purl.obolibrary.org/obo/PATO_0000122
-%  source ontology: PATO
-%  entity: http://purl.obolibrary.org/obo/PATO_0000122
-%  label: pressure
-%  parent: http://purl.obolibrary.org/obo/BFO_0000019
-%  parentLabel: quality
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (pressure(x) -> quality(x)). 
-
