--- Scrub_Team_2018/FOL_Output/SIO_output.txt
@@ -1,11 +1,0 @@
-%  http://semanticscience.org/resource/SIO_000184
-%  source ontology: SIO
-%  entity: http://semanticscience.org/resource/SIO_000184
-%  label: legal name
-%  parent: http://purl.obolibrary.org/obo/IAO_0000302
-%  parentLabel: author identification
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (legal_name(x) -> author_identification(x)). 
-
