--- Scrub_Team_2018/FOL_Output/UO_output.txt
@@ -1,99 +1,0 @@
-%  http://purl.obolibrary.org/obo/UO_0000003
-%  source ontology: UO
-%  entity: http://purl.obolibrary.org/obo/UO_0000003
-%  label: length unit
-%  parent: http://purl.obolibrary.org/obo/IAO_0000003
-%  parentLabel: measurement unit label
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (length_unit(x) -> measurement_unit_label(x)). 
-
-%  http://purl.obolibrary.org/obo/UO_0000003
-%  source ontology: UO
-%  entity: http://purl.obolibrary.org/obo/UO_0000003
-%  label: mass unit
-%  parent: http://purl.obolibrary.org/obo/IAO_0000003
-%  parentLabel: measurement unit label
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (mass_unit(x) -> measurement_unit_label(x)). 
-
-%  http://purl.obolibrary.org/obo/UO_0000003
-%  source ontology: UO
-%  entity: http://purl.obolibrary.org/obo/UO_0000003
-%  label: time unit
-%  parent: http://purl.obolibrary.org/obo/IAO_0000003
-%  parentLabel: measurement unit label
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (time_unit(x) -> measurement_unit_label(x)). 
-
-%  http://purl.obolibrary.org/obo/UO_0000002
-%  source ontology: UO
-%  entity: http://purl.obolibrary.org/obo/UO_0000002
-%  label: length unit
-%  parent: http://purl.obolibrary.org/obo/IAO_0000003
-%  parentLabel: measurement unit label
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (length_unit(x) -> measurement_unit_label(x)). 
-
-%  http://purl.obolibrary.org/obo/UO_0000002
-%  source ontology: UO
-%  entity: http://purl.obolibrary.org/obo/UO_0000002
-%  label: mass unit
-%  parent: http://purl.obolibrary.org/obo/IAO_0000003
-%  parentLabel: measurement unit label
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (mass_unit(x) -> measurement_unit_label(x)). 
-
-%  http://purl.obolibrary.org/obo/UO_0000002
-%  source ontology: UO
-%  entity: http://purl.obolibrary.org/obo/UO_0000002
-%  label: time unit
-%  parent: http://purl.obolibrary.org/obo/IAO_0000003
-%  parentLabel: measurement unit label
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (time_unit(x) -> measurement_unit_label(x)). 
-
-%  http://purl.obolibrary.org/obo/UO_0000001
-%  source ontology: UO
-%  entity: http://purl.obolibrary.org/obo/UO_0000001
-%  label: length unit
-%  parent: http://purl.obolibrary.org/obo/IAO_0000003
-%  parentLabel: measurement unit label
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (length_unit(x) -> measurement_unit_label(x)). 
-
-%  http://purl.obolibrary.org/obo/UO_0000001
-%  source ontology: UO
-%  entity: http://purl.obolibrary.org/obo/UO_0000001
-%  label: mass unit
-%  parent: http://purl.obolibrary.org/obo/IAO_0000003
-%  parentLabel: measurement unit label
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (mass_unit(x) -> measurement_unit_label(x)). 
-
-%  http://purl.obolibrary.org/obo/UO_0000001
-%  source ontology: UO
-%  entity: http://purl.obolibrary.org/obo/UO_0000001
-%  label: time unit
-%  parent: http://purl.obolibrary.org/obo/IAO_0000003
-%  parentLabel: measurement unit label
-%  EquivalenceAxiom: None
-%  SubclassAxiom: None
-
-all x (time_unit(x) -> measurement_unit_label(x)). 
-
