--- Scrub_Team_2018/to_do_list.txt
@@ -1,37 +1,0 @@
-List of to-dos for ICO and RUBRIC
-
-Up Next:
-- RUBRIC: 'Answer directive' label
-- RUBRIC: Every conditional will have a trigger condition description ('if') and a consequent directive ('then').
-- Medical procedure should be 'research procedure' and 'clinical procedure'
-- Planned and Actual study duration measurements are different
-- Different states for informed consent forms
-  - pre-approval vs approved vs filled out (where does each class fall?)
-- Informing does not have to be human-human. I can learn from a book. New class?
-- Import RO temporal relations
-- add RO to imports
-- RUBRIC: specimen use/data use
-- Are these social acts?
-  - 'canceling a permission'
-  - 'declining to provide consent'
-- 'consenter' and 'agent' not the same superclass (per JV's email)?
-
-FOL script
-- remove comments and rerun
-- single comment: human readable comment for each axiom
-- run query in protege and retrieve (1 hour)
-- camelcase instead of underscore
-
-Final:
-- RUBRIC: For all CCO terms, add the license info in an annotation property
-- Spell check all annotations
-- Grammar check all annotations
-- All definitions have 'genus - differentia' form
-- Protege reasoner to check axiomatic consistency
-- Resolve synonyms:
-  - Known issues: permissions/permitee, subject/person/agent/patient
-- FOL is consistent
-- review all obsolete classes
-- new PURLS resolve in OBO
-- 'research study' should be replaced by OBI: investigation
-- create ontology publishing process
